+++ tests/testsrc/commontest.cpp
@@ -7,19 +7,19 @@
 #include "common.h"
 
 TEST(Common, HeightmapFromArray){
-    int32_t map_array[9] = {
-        1, 1, 1,
-        1, 3, 1,
-        1, 1, 1
+    float map_array[9] = {
+        1.0, 1.0, 1.0,
+        1.0, 3.0, 1.0,
+        1.0, 1.0, 1.0
     };
 
     vs_heightmap_t map = heightmap_from_array(3, 3, map_array);
     
-    EXPECT_EQ(map.cols, 3);
-    EXPECT_EQ(map.rows, 3);
+    EXPECT_EQ(map.cols, 3.0);
+    EXPECT_EQ(map.rows, 3.0);
 
-    EXPECT_EQ(map.heightmap[1*map.cols + 1], 3);
-    EXPECT_EQ(map.heightmap[0*map.cols + 0], 1);
+    EXPECT_EQ(map.heightmap[1*map.cols + 1], 3.0);
+    EXPECT_EQ(map.heightmap[0*map.cols + 0], 1.0);
 }
 
 TEST(Common, HeightmapFromFile){
